+++ android/gradle.properties
@@ -4,6 +4,7 @@
 # Gradle settings configured through the IDE *will override*
 # any settings specified in this file.
 
+
 # For more details on how to configure your build environment visit
 # http://www.gradle.org/docs/current/userguide/build_environment.html
